# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Clinic Forms Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
        - name: archived
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get an invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Archive an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Archived
  /invoices/{id}/restore:
    patch:
      operationId: restoreInvoice
      tags: [invoices]
      summary: Restore an archived invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

  # ─── CUSTOMERS ────────────────────────────────────────────────────────────────
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DOCTORS ──────────────────────────────────────────────────────────────────
  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: List all doctors
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"
    post:
      operationId: createDoctor
      tags: [doctors]
      summary: Create a doctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
  /doctors/{id}:
    get:
      operationId: getDoctor
      tags: [doctors]
      summary: Get a doctor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
        "404":
          description: Not found
    patch:
      operationId: updateDoctor
      tags: [doctors]
      summary: Update a doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
    delete:
      operationId: deleteDoctor
      tags: [doctors]
      summary: Delete a doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── USERS ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Invoice:
      type: object
      required: [id, invoiceNumber, invoiceDate, totalAmount, isArchived, createdAt, items]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        invoiceDate:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerCode:
          type: ["string", "null"]
        branch:
          type: ["string", "null"]
        section:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        collector:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        subtotal:
          type: number
        discount:
          type: number
        totalAmount:
          type: number
        isArchived:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
    InvoiceItem:
      type: object
      required: [id, invoiceId, itemCode, itemName, unit, quantity, price, total]
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        itemCode:
          type: string
        itemName:
          type: string
        unit:
          type: string
        quantity:
          type: number
        bonus:
          type: number
        price:
          type: number
        total:
          type: number
    InvoiceInput:
      type: object
      required: [invoiceDate, items]
      properties:
        invoiceDate:
          type: string
        customerId:
          type: integer
        branch:
          type: string
        section:
          type: string
        department:
          type: string
        collector:
          type: string
        notes:
          type: string
        discount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItemInput"
    InvoiceItemInput:
      type: object
      required: [itemCode, itemName, unit, quantity, price]
      properties:
        itemCode:
          type: string
        itemName:
          type: string
        unit:
          type: string
        quantity:
          type: number
        bonus:
          type: number
        price:
          type: number
    InvoiceUpdate:
      type: object
      properties:
        invoiceDate:
          type: string
        customerId:
          type: integer
        branch:
          type: string
        section:
          type: string
        department:
          type: string
        collector:
          type: string
        notes:
          type: string
        discount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItemInput"

    # ─── CUSTOMER ───────────────────────────────────────────────────────────────
    Customer:
      type: object
      required: [id, customerCode, name, createdAt]
      properties:
        id:
          type: integer
        customerCode:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    CustomerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string

    # ─── DOCTOR ─────────────────────────────────────────────────────────────────
    Doctor:
      type: object
      required: [id, name, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        specialty:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    DoctorInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        specialty:
          type: string
    DoctorUpdate:
      type: object
      properties:
        name:
          type: string
        specialty:
          type: string
        isActive:
          type: boolean

    # ─── USER ───────────────────────────────────────────────────────────────────
...