# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Clinic Forms Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerUpdate {
  name?: string;
  phone?: string;
  address?: string;
  notes?: string;
}